reorganize namespaces
