+++ mani/domain/testing/testing_domain_module.py
@@ -2,8 +2,8 @@ from typing import List, Type, Tuple
 
 from injector import Module, Scope, inject, SingletonScope, Binder
 
-from common.plain_effect_store import PlainEffectStore
-from common.test_cases.save_effects_bus_hook import SaveEffectsBusHook
+from mani.infrastructure.testing.plain_effect_store import PlainEffectStore
+from mani.domain.testing.save_effects_bus_hook import SaveEffectsBusHook
 from mani.domain.cqrs.effect_store.effect_store import EffectStore
 from mani.domain.cqrs.effects import Command
 from mani.domain.model.modules import DomainModule
