+++ mani/infrastructure/testing/plain_effect_store.py
